Move RedirectToRouteRector into Symfony26 category
