+++ config/sets/symfony/symfony26.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Symfony\Rector\MethodCall\RedirectToRouteRector;
+use Rector\Symfony\Symfony26\Rector\MethodCall\RedirectToRouteRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(RedirectToRouteRector::class);
