+++ rules-tests/Symfony26/Rector/MethodCall/RedirectToRouteRector/Fixture/fixture3.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Symfony\Tests\Rector\MethodCall\RedirectToRouteRector\Fixture;
+namespace Rector\Symfony\Tests\Symfony26\Rector\MethodCall\RedirectToRouteRector\Fixture;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
@@ -16,7 +16,7 @@ class AppController3 extends Controller
 -----
 <?php
 
-namespace Rector\Symfony\Tests\Rector\MethodCall\RedirectToRouteRector\Fixture;
+namespace Rector\Symfony\Tests\Symfony26\Rector\MethodCall\RedirectToRouteRector\Fixture;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
