+++ rules-tests/Symfony26/Rector/MethodCall/RedirectToRouteRector/Fixture/generate_url_with_relative.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Symfony\Tests\Rector\MethodCall\RedirectToRouteRector\Fixture;
+namespace Rector\Symfony\Tests\Symfony26\Rector\MethodCall\RedirectToRouteRector\Fixture;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
