+++ rules-tests/Symfony26/Rector/MethodCall/RedirectToRouteRector/Fixture/skip_other_generator.php.inc
@@ -1,7 +1,8 @@
 <?php
 
-namespace Rector\Symfony\Tests\Rector\MethodCall\RedirectToRouteRector\Fixture;
+namespace Rector\Symfony\Tests\Symfony26\Rector\MethodCall\RedirectToRouteRector\Fixture;
 
+use Rector\Symfony\Tests\Rector\MethodCall\RedirectToRouteRector\Fixture\SomeGenerator;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class SkipController extends AbstractController
