+++ rules-tests/Symfony26/Rector/MethodCall/RedirectToRouteRector/config/configured_rule.php
@@ -3,9 +3,9 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Symfony\Rector\MethodCall\RedirectToRouteRector;
+use Rector\Symfony\Symfony26\Rector\MethodCall\RedirectToRouteRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
+    $rectorConfig->import(__DIR__ . '/../../../../../../config/config.php');
     $rectorConfig->rule(RedirectToRouteRector::class);
 };
